BUGFIX/MAJOR(keystone): Fix invalid logrotation parameters
 ##### SUMMARY

The current logrotate configuration causes a recreation of file as root:keystone with 600 permissions, leading the keystone service to go in error (because of a permission denied on the log file), returning 500 to all requests.

It is possible that this is a race-condition during log rotation between keystone (that creates a file with keystone:keystone 0600), and logrotate.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION
OS-496
Link: https://github.com/openstack/rpm-packaging/blob/master/openstack/keystone/openstack-keystone.logrotate